@@ -10,120 +10,92 @@ class DepartamentoController extends Controller
 {
     /**
      * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
      */
     public function index()
     {
         $departamentos = DB::table('tb_departamento')
-        ->join('tb_pais', 'tb_departamento.pais_codi', '=' , 'tb_pais.pais_codi')
-        ->select('tb_departamento.*' , "tb_pais.pais_nomb")
+        ->join('tb_pais', 'tb_departamento.pais_codi', '=', 'tb_pais.pais_codi')
+        ->select('tb_departamento.*', 'tb_pais.pais_nomb')
         ->get();
-     
-     return view('departamento.index', ['departamentos' => $departamentos]);
+        return view('departamento.index', ['departamentos'=>$departamentos]);
     }
 
     /**
      * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
      */
     public function create()
     {
-        $paises = DB::table('tb_pais')
-            ->orderBy('pais_nomb')
-            ->get();
-        return view('departamento.new', ['paises' => $paises]);
+        $pais = DB::table('tb_pais')
+        ->orderBy('pais_nomb')
+        ->get();
+        return view('departamento.new', ['pais' => $pais]);
     }
 
     /**
      * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         $departamento = new Departamento();
-
+        // $departamento->depa_codi = $request->id;
+        // El código de departamento es auto incremental
         $departamento->depa_nomb = $request->name;
         $departamento->pais_codi = $request->code;
         $departamento->save();
-
         $departamentos = DB::table('tb_departamento')
-        ->join('tb_pais', 'tb_departamento.pais_codi', '=' , 'tb_pais.pais_codi')
-        ->select('tb_departamento.*' , "tb_pais.pais_nomb")
+        ->join('tb_pais', 'tb_departamento.pais_codi', '=', 'tb_pais.pais_codi')
+        ->select('tb_departamento.*', 'tb_pais.pais_nomb')
         ->get();
-
         return view('departamento.index', ['departamentos' => $departamentos]);
     }
 
     /**
      * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(string $id)
     {
         //
     }
 
     /**
      * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(string $id)
     {
-        $departamento = Departamento::find($id);
-        $paises = DB::table('tb_pais')
+        $departamento= Departamento::find($id);
+        $pais=DB::table('tb_pais')
         ->orderBy('pais_nomb')
         ->get();
-
-        return view ('departamento.edit' , ['departamento' => $departamento, 'paises' => $paises]);
+        return view('departamento.edit', ['departamento' => $departamento, 'pais'=>$pais]);
     }
 
     /**
      * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, string $id)
     {
-        $departamento = Departamento::find($id);
-
+        $departamento= Departamento::find($id);
         $departamento->depa_nomb = $request->name;
         $departamento->pais_codi = $request->code;
         $departamento->save();
-
-        $departamentos = DB::table('tb_departamento')
-        ->join('tb_pais', 'tb_departamento.pais_codi', '=' , 'tb_pais.pais_codi')
+        $departamentos=DB::table('tb_departamento')
+        ->join('tb_pais', 'tb_departamento.pais_codi', '=', 'tb_pais.pais_codi')
         ->select('tb_departamento.*', 'tb_pais.pais_nomb')
         ->get();
-
-        return view ('departamento.index' , ['departamentos' => $departamentos]);
+        return view('departamento.index', ['departamentos' => $departamentos]);
     }
 
     /**
      * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(string $id)
     {
-        $departamento = Departamento::find($id);
+        $departamento= Departamento::find($id);
         $departamento->delete();
-
-        $departamentos = DB::table('tb_departamento')
-        ->join('tb_pais', 'tb_departamento.pais_codi', '=' , 'tb_pais.pais_codi')
+        $departamentos=DB::table('tb_departamento')
+        ->join('tb_pais', 'tb_departamento.pais_codi', '=', 'tb_pais.pais_codi')
         ->select('tb_departamento.*', 'tb_pais.pais_nomb')
         ->get();
-
-        return view ('departamento.index' , ['departamentos' => $departamentos]);
+        return view('departamento.index', ['departamentos' => $departamentos]);
     }
 }
