@@ -8,7 +8,7 @@ use Illuminate\Database\Eloquent\Model;
 class Comuna extends Model
 {
     use HasFactory;
-    protected $table = 'tb_comuna';
-    protected $primaryKey = 'comu_codi';
+    protected $table = 'tb_comuna' ;
+    protected $primaryKey = 'comu_codi' ;
     public $timestamps = false;
 }
