@@ -8,7 +8,7 @@ use Illuminate\Database\Eloquent\Model;
 class Municipio extends Model
 {
     use HasFactory;
-    protected $table = 'tb_municipio';
-    protected $primaryKey = 'muni_codi';
+    protected $table = 'tb_municipio' ;
+    protected $primaryKey = 'muni_codi' ;
     public $timestamps = false;
 }
